# docs/swagger.yaml
definitions:
...
  consts.DictItemStatus:
    enum:
    - 0
    type: integer
    x-enum-comments:
      DictItemStatusEnable: 启用
    x-enum-varnames:
    - DictItemStatusEnable
...
  consts.MenuStatus:
    enum:
    - 0
    type: integer
    x-enum-comments:
      MenuStatusEnable: 启用
    x-enum-varnames:
    - MenuStatusEnable
  consts.MenuType:
    enum:
    - 0
    - 1
    type: integer
    x-enum-comments:
      MenuTypeCatalog: 目录
      MenuTypeMenu: 菜单
    x-enum-varnames:
    - MenuTypeCatalog
    - MenuTypeMenu
...
  dto.SysDictItemAddDto:
    properties:
      dictId:
        description: 字典id
        type: integer
      dictItemName:
        description: 选项名称
        type: string
      dictItemValue:
        description: 选项值
        type: string
      remark:
        description: 备注
        type: string
      status:
        allOf:
        - $ref: '#/definitions/consts.DictItemStatus'
        description: 配置状态（0正常 1停用）
    type: object
  dto.SysDictItemUpdateDto:
    properties:
      dictId:
        description: 字典id
        type: integer
      dictItemName:
        description: 选项名称
        type: string
      dictItemValue:
        description: 选项值
        type: string
      id:
        description: 编号
        type: integer
      remark:
        description: 备注
        type: string
      status:
        allOf:
        - $ref: '#/definitions/consts.DictItemStatus'
        description: 配置状态（0正常 1停用）
    type: object
  dto.SysDictItemVo:
    properties:
      createdAt:
        description: 创建时间
        type: string
      dictId:
        description: 字典id
        type: integer
      dictItemName:
        description: 选项名称
        type: string
      dictItemValue:
        description: 选项值
        type: string
      id:
        description: 编号
        type: integer
      remark:
        description: 备注
        type: string
      status:
        allOf:
        - $ref: '#/definitions/consts.DictItemStatus'
        description: 配置状态（0正常 1停用）
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  dto.SysGenTableAddDto:
    properties:
      businessName:
        description: 业务名称
        type: string
      menuParentId:
        description: 上级菜单
        type: integer
      modelName:
        description: 实体名称
        type: string
      moduleName:
        description: 模块名称
        type: string
      tableDescription:
        description: 表描述
        type: string
      tableName:
        description: 表名称
        type: string
    type: object
  dto.SysGenTableColumnVo:
    properties:
      codeField:
        description: 代码字段
        type: string
      codeType:
        description: 代码类型
        type: string
      columnDescription:
        description: 字段描述
        type: string
      columnName:
        description: 字段名
        type: string
      columnType:
        description: 字段类型
        type: string
      componentType:
        description: 组件类型
        type: string
      createdAt:
        description: 创建时间
        type: string
      dictCode:
        description: 字典代码
        type: string
      id:
        description: Id
        type: integer
      isEdit:
        description: 是否编辑字段（0否 1是）
        type: boolean
      isList:
        description: 是否列表字段（0否 1是）
        type: boolean
      isPk:
        description: 是否主键（0否 1是）
        type: boolean
      isQuery:
        description: 是否查询字段（0否 1是）
        type: boolean
      paramName:
        description: 参数名称
        type: string
      queryMethod:
        description: 查询方式
        type: string
      tableId:
        description: 表编号
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  dto.SysGenTableUpdateDto:
    properties:
      businessName:
        description: 业务名称
        type: string
      id:
        description: 编号
        type: integer
      menuParentId:
        description: 上级菜单
        type: integer
      modelName:
        description: 实体名称
        type: string
      moduleName:
        description: 模块名称
        type: string
      tableDescription:
        description: 表描述
        type: string
      tableName:
        description: 表名称
        type: string
    type: object
  dto.SysGenTableVo:
    properties:
      businessName:
        description: 业务名称
        type: string
      columnList:
        description: 字段列表
        items:
          $ref: '#/definitions/dto.SysGenTableColumnVo'
        type: array
      createdAt:
        description: 创建时间
        type: string
      id:
        type: integer
      menuParentId:
        description: 上级菜单
        type: integer
      modelName:
        description: 实体名称
        type: string
      moduleAlias:
        description: 模块别名
        type: string
      moduleName:
        description: 模块名称
        type: string
      tableDescription:
        description: 表描述
        type: string
      tableName:
        description: 表名称
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  dto.SysLoginLogVo:
    properties:
      browser:
        description: 浏览器
        type: string
      id:
        description: 字典id
        type: integer
      ipAddress:
        description: IP地址
        type: string
      location:
        description: 登录位置
        type: string
      loginTime:
        description: 登录时间
        type: string
      operatingSystem:
        description: 操作系统
        type: string
      userId:
        description: 用户id
        type: integer
    type: object
  dto.SysMenuAddDto:
    properties:
      component:
        description: 组件路径
        type: string
      icon:
        description: 菜单图标
        type: string
      isAffix:
        description: 是否固定tagsView
        type: boolean
      isCache:
        description: 是否缓存
        type: boolean
      isHide:
        description: 是否隐藏
        type: boolean
      isIframe:
        description: 是否内嵌
        type: boolean
      link:
        description: 外部地址
        type: string
      menuName:
        description: 菜单名称
        type: string
      menuType:
        allOf:
        - $ref: '#/definitions/consts.MenuType'
        description: 菜单类型（0菜单，1按钮）
      parentId:
        description: 父菜单ID
        type: integer
      path:
        description: 路由地址
        type: string
      permission:
        description: 按钮权限
        type: string
      remark:
        description: 备注
        type: string
      requestMethod:
        description: 请求方法
        type: string
      sort:
        description: 排序
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/consts.MenuStatus'
        description: 菜单状态（0正常 1停用）
    type: object
  dto.SysMenuUpdateDto:
    properties:
      component:
        description: 组件路径
        type: string
      icon:
        description: 菜单图标
        type: string
      id:
        description: 编号
        type: integer
      isAffix:
        description: 是否固定tagsView
        type: boolean
      isCache:
        description: 是否缓存
        type: boolean
      isHide:
        description: 是否隐藏
        type: boolean
      isIframe:
        description: 是否内嵌
        type: boolean
      link:
        description: 外部地址
        type: string
      menuName:
        description: 菜单名称
        type: string
      menuType:
        allOf:
        - $ref: '#/definitions/consts.MenuType'
        description: 菜单类型（0菜单，1按钮）
      parentId:
        description: 父菜单ID
        type: integer
      path:
        description: 路由地址
        type: string
      permission:
        description: 按钮权限
        type: string
      remark:
        description: 备注
        type: string
      requestMethod:
        description: 请求方法
        type: string
      sort:
        description: 排序
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/consts.MenuStatus'
        description: 菜单状态（0正常 1停用）
    type: object
  dto.SysMenuVo:
    properties:
      children:
        description: 子菜单
        items:
          $ref: '#/definitions/dto.SysMenuVo'
        type: array
      component:
        description: 组件路径
        type: string
      createdAt:
        description: 创建时间
        type: string
      icon:
        description: 菜单图标
        type: string
      id:
        description: 编号
        type: integer
      isAffix:
        description: 是否固定tagsView
        type: boolean
      isCache:
        description: 是否缓存
        type: boolean
      isHide:
        description: 是否隐藏
        type: boolean
      isIframe:
        description: 是否内嵌
        type: boolean
      link:
        description: 外部地址
        type: string
      menuName:
        description: 菜单名称
        type: string
      menuType:
        allOf:
        - $ref: '#/definitions/consts.MenuType'
        description: 菜单类型（0菜单，1按钮）
      parentId:
        description: 父菜单ID
        type: integer
      path:
        description: 路由地址
        type: string
      permission:
        description: 按钮权限
        type: string
      remark:
        description: 备注
        type: string
      requestMethod:
        description: 请求方法
        type: string
      sort:
        description: 排序
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/consts.MenuStatus'
        description: 菜单状态（0正常 1停用）
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  dto.TableInfoVo:
    properties:
      businessName:
        description: 业务名称
        type: string
      modelName:
        description: 实体名称
        type: string
      moduleName:
        description: 模块名称
        type: string
      tableDescription:
        description: 表描述
        type: string
      tableName:
        description: 表名称
        type: string
    type: object
...
  request.IdInfoDto:
    properties:
      id:
        description: id编号
        type: integer
    type: object
  response.JsonResult:
    properties:
      code:
        description: 状态码（架构层面：中间件/异常处理等）
        type: integer
      data:
        description: 数据
      message:
        description: 消息
        type: string
      success:
        description: 是否成功（业务层面）
        type: boolean
    type: object
  response.PageResult:
    properties:
      list:
        description: 数据
      totalCount:
        description: 总条数
        type: integer
    type: object
info:
  contact: {}
  description: go-fast-admin 接口文档
  title: 系统模块
  version: 1.0.0
paths:
...
  /system/dictItem/add:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysDictItemAddDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 添加字典选项
      tags:
      - 字典选项
  /system/dictItem/delete:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdInfoDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 删除字典选项
      tags:
      - 字典选项
  /system/dictItem/detail:
    get:
      parameters:
      - description: id编号
        in: query
        name: id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.SysDictItemVo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取字典选项详情
      tags:
      - 字典选项
  /system/dictItem/list:
    get:
      parameters:
      - description: 字典id
        in: query
        name: dictId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SysDictItemVo'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 字典选项列表
      tags:
      - 字典选项
  /system/dictItem/query:
    get:
      parameters:
      - description: 页码
        in: query
        name: pageIndex
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.SysDictItemVo'
                        type: array
                    type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 字典选项查询
      tags:
      - 字典选项
  /system/dictItem/update:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysDictItemUpdateDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 更新字典选项
      tags:
      - 字典选项
  /system/genTable/add:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysGenTableAddDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 添加代码生成（表）
      tags:
      - 代码生成（表）
  /system/genTable/delete:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdInfoDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 删除代码生成（表）
      tags:
      - 代码生成（表）
  /system/genTable/detail:
    get:
      parameters:
      - description: id编号
        in: query
        name: id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.SysGenTableVo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取代码生成（表）详情
      tags:
      - 代码生成（表）
  /system/genTable/previewCode:
    get:
      parameters:
      - description: 表id
        in: query
        name: tableId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 预览代码
      tags:
      - 代码生成（表）
  /system/genTable/query:
    get:
      parameters:
      - description: 页码
        in: query
        name: pageIndex
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 表名称
        in: query
        name: tableName
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.SysGenTableVo'
                        type: array
                    type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 代码生成（表）查询
      tags:
      - 代码生成（表）
  /system/genTable/tableList:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.TableInfoVo'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取表列表
      tags:
      - 代码生成（表）
  /system/genTable/update:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysGenTableUpdateDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 更新代码生成（表）
      tags:
      - 代码生成（表）
  /system/genTableColumn/list:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SysGenTableColumnVo'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 代码生成（表字段）列表
      tags:
      - 代码生成（表字段）
  /system/genTableColumn/update:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          items:
            $ref: '#/definitions/dto.SysGenTableUpdateDto'
          type: array
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 更新代码生成（表字段）
      tags:
      - 代码生成（表字段）
...
  /system/loginLog/detail:
    get:
      parameters:
      - description: id编号
        in: query
        name: id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.SysLoginLogVo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取登录日志详情
      tags:
      - 登录日志
...
  /system/loginLog/query:
    get:
      parameters:
      - description: 页码
        in: query
        name: pageIndex
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.SysLoginLogVo'
                        type: array
                    type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 登录日志查询
      tags:
      - 登录日志
...
  /system/menu/add:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysMenuAddDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 添加菜单
      tags:
      - 菜单
  /system/menu/delete:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdInfoDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 删除菜单
      tags:
      - 菜单
  /system/menu/detail:
    get:
      parameters:
      - description: id编号
        in: query
        name: id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.SysMenuVo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取菜单详情
      tags:
      - 菜单
...
  /system/menu/query:
    get:
      parameters:
      - description: 菜单名称
        in: query
        name: menuName
        type: string
      - description: 页码
        in: query
        name: pageIndex
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.SysMenuVo'
                        type: array
                    type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 菜单查询
      tags:
      - 菜单
  /system/menu/tree:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SysMenuVo'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 菜单树状
      tags:
      - 菜单
  /system/menu/update:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysMenuUpdateDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 更新菜单
      tags:
      - 菜单
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"